map(): check for "Canceled" field
Problem: As mentioned in #2, some of the rows in
customer-reservations.csv contain a field indicating the status of the
reservation, namely if the reservation has been canceled or not. If it
has been canceled, its revenue should not be added to the total revenue
for a given month-year composite key.

Add logic to the map () function that checks the "booking_status" field
and set avg_price_per_room to 0.0 if the reservation was canceled.