map (): check booking_status for hotel-bookings.csv
The hotel-bookings.csv file also has a field that indictates whether a
customer canceled their reservation; if they did, their revenue
shouldn't be added to the month/year's total revenue.

Add logic in the map () function that checks the status of the
reservation and set avg_price_per_room to 0.0 if the reservation was
canceled.