map (): calculate total cost of reservation
The map () function currently just outputs the average price per room
and not the total cost of each reservation.

Add logic to the map () function to calculate how many nights each
reservation was for and then multiply it by the average price per room
to get a total cost. Set the total cost as the value for each month/year
key to be sent to reduce ().